+++ packages/cinterop/src/jvm/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -17,8 +17,6 @@
 package io.realm.internal.interop
 
 import io.realm.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
-import io.realm.internal.interop.RealmInterop.asObjectId
-import io.realm.internal.interop.RealmInterop.cptr
 import io.realm.internal.interop.sync.AuthProvider
 import io.realm.internal.interop.sync.CoreSubscriptionSetState
 import io.realm.internal.interop.sync.CoreUserState
@@ -378,7 +376,7 @@ actual object RealmInterop {
 
     private fun from_realm_value(value: realm_value_t): RealmValue {
         return RealmValue(
-            when (value?.type) {
+            when (value.type) {
                 realm_value_type_e.RLM_TYPE_STRING ->
                     value.string
                 realm_value_type_e.RLM_TYPE_INT ->
@@ -408,6 +406,10 @@ actual object RealmInterop {
         realmc.realm_set_value(obj.cptr(), key.key, cvalue, isDefault)
     }
 
+    actual fun realm_set_embedded(obj: RealmObjectPointer, key: PropertyKey): RealmObjectPointer {
+        return LongPointerWrapper(realmc.realm_set_embedded(obj.cptr(), key.key))
+    }
+
     actual fun realm_get_list(obj: RealmObjectPointer, key: PropertyKey): RealmListPointer {
         return LongPointerWrapper(
             realmc.realm_get_list(
@@ -434,10 +436,22 @@ actual object RealmInterop {
         realmc.realm_list_insert(list.cptr(), index, cvalue)
     }
 
+    actual fun realm_list_insert_embedded(list: RealmListPointer, index: Long): RealmObjectPointer {
+        return LongPointerWrapper(realmc.realm_list_insert_embedded(list.cptr(), index))
+    }
+
     actual fun realm_list_set(list: RealmListPointer, index: Long, value: RealmValue): RealmValue {
         return realm_list_get(list, index).also {
             realmc.realm_list_set(list.cptr(), index, to_realm_value(value))
         }
+    }
+
+    actual fun realm_list_set_embedded(list: RealmListPointer, index: Long): RealmValue {
+        // Returns the new object as a Link to follow convention of other getters and allow to
+        // reuse the converter infrastructure
+        val embedded = realmc.realm_list_set_embedded(list.cptr(), index)
+        val link = realmc.realm_object_as_link(embedded)
+        return RealmValue(Link(ClassKey(link.target_table), link.target))
     }
 
     actual fun realm_list_clear(list: RealmListPointer) {