+++ packages/library-base/src/commonMain/kotlin/io/realm/schema/RealmStorageType.kt
@@ -17,8 +17,10 @@
 package io.realm.schema
 
 import io.realm.BaseRealmObject
+import io.realm.EmbeddedRealmObject
 import io.realm.ObjectId
 import io.realm.RealmInstant
+import io.realm.RealmObject
 import kotlin.reflect.KClass
 
 /**
@@ -43,7 +45,7 @@ public enum class RealmStorageType(public val kClass: KClass<*>) {
     STRING(String::class),
 
     /**
-     * Storage type for properties of type [RealmObject].
+     * Storage type for properties of type [RealmObject] or [EmbeddedRealmObject].
      */
     OBJECT(BaseRealmObject::class),